Separate parent routes from main

+++ routes/card.js
@@ -1,13 +1,15 @@
 const express = require('express');
 const router = express.Router();
+const flashMessage = require('../helpers/messenger');
 const moment = require('moment');
 const Card = require('../models/Card');
-const flashMessage = require('../helpers/messenger');
+const ensureAuthenticated = require('../helpers/auth');
 
 
 // get --> List of Cards
-router.get('/listCard', (req, res) => {
+router.get('/listCard', ensureAuthenticated.ensureParent, (req, res) => {
     Card.findAll({
+        where: { parentTutorId: req.user.id },
         order: [['expiryDate', 'DESC']],
         raw: true
     })
@@ -19,39 +21,37 @@ router.get('/listCard', (req, res) => {
 
 
 // get --> Add Card
-router.get('/addCard', (req, res) => {
-    async function hello() {                                        // check if COUNT(cards) >= 3
-        let isOk = true;
-        const myCount = await Card.count();
-        console.log("Number of cards: " + myCount);                   
-        if (myCount >= 3) {                                         // yes --> redirect to card/listCard
-            flashMessage(res, 'error', 'Maximum number of cards');
-            isOk = false;
-        }
-        if (!isOk) {
-            Card.findAll({
-                order: [['expiryDate', 'DESC']],
-                raw: true
-            })
-                .then((cards) => {
-                    res.render('card/listCard', { cards });
-                })
-                .catch(err => console.log(err));
-            return;
-        }                       
-        res.render('card/addCard');                                 // no --> render card/addCard
+router.get('/addCard', ensureAuthenticated.ensureParent, async function (req, res) {                                      
+    let isOk = true;
+    const myCount = await Card.count();                         // check if COUNT(cards) >= 3
+    console.log("Number of cards: " + myCount);                   
+    if (myCount >= 3) {                                         // yes --> redirect to card/listCard
+        flashMessage(res, 'error', 'Maximum number of cards');
+        isOk = false;
     }
-    hello();
+    if (!isOk) {
+        Card.findAll({
+            order: [['expiryDate', 'DESC']],
+            raw: true
+        })
+            .then((cards) => {
+                res.render('card/listCard', { cards });
+            })
+            .catch(err => console.log(err));
+        return;
+    }                       
+    res.render('card/addCard');                                 // no --> render card/addCard
 });
 
 
 // post --> Add Card
-router.post('/addCard', (req, res) => {
+router.post('/addCard', ensureAuthenticated.ensureParent,  (req, res) => {
     let cardName = req.body.cardName;
     let cardNo = req.body.cardNo.slice(0, 19);
     let expiryDate = moment(req.body.expiryDate, 'MM/YYYY');
     let ccv = req.body.ccv;
     let debitCredit = req.body.debitCredit;
+    let parentTutorId = req.user.id;
     let isValid = true;
     if (cardNo.length < 13) {
         flashMessage(res, 'error', 'Card Number must contain 13 to 19 numbers (inclusive).');
@@ -75,7 +75,7 @@ router.post('/addCard', (req, res) => {
         });
         return;
     }
-    Card.create({ cardName, cardNo, expiryDate, ccv, debitCredit })
+    Card.create({ cardName, cardNo, expiryDate, ccv, debitCredit, parentTutorId })
         .then((card) => {
             console.log(card.toJSON());
             flashMessage(res, 'success', cardName + ' ' + debitCredit + ' card saved successfully');
@@ -86,14 +86,19 @@ router.post('/addCard', (req, res) => {
 
 
 // get --> Edit Card
-router.get('/editCard/:id', (req, res) => {
+router.get('/editCard/:id', ensureAuthenticated.ensureParent,  (req, res) => {
     Card.findByPk(req.params.id)
         .then((card) => {
             if (!card) {
                 flashMessage(res, 'error', 'Payment Card not found');
                 res.redirect('/card/listCard');
                 return;
-                }
+            }
+            if (req.user.id != card.parentTutorId) {
+                flashMessage(res, 'error', 'Unauthorised access');
+                res.redirect('/card/listCard');
+                return;
+            }
             res.render('card/editCard', { card });
         })
         .catch(err => console.log(err));
@@ -101,7 +106,7 @@ router.get('/editCard/:id', (req, res) => {
 
 
 // post --> Edit Card
-router.post('/editCard/:id', (req, res) => {
+router.post('/editCard/:id', ensureAuthenticated.ensureParent,  (req, res) => {
     let cardName = req.body.cardName;
     let cardNo = req.body.cardNo.slice(0, 19);
     let expiryDate = moment(req.body.expiryDate, 'MM/YYYY');
@@ -146,7 +151,7 @@ router.post('/editCard/:id', (req, res) => {
 
 
 // get --> Delete Card
-router.get('/deleteCard/:id', async function (req, res) {
+router.get('/deleteCard/:id', ensureAuthenticated.ensureParent, async function (req, res) {
     try {
         let card = await Card.findByPk(req.params.id);
         if (!card) {
@@ -154,11 +159,11 @@ router.get('/deleteCard/:id', async function (req, res) {
             res.redirect('/card/listCard');
             return;
         }
-        // if (req.user.id != video.userId) {
-        //     flashMessage(res, 'error', 'Unauthorised access');
-        //     res.redirect('/video/listVideos');
-        //     return;
-        // }
+        if (req.user.id != card.parentTutorId) {
+            flashMessage(res, 'error', 'Unauthorised access');
+            res.redirect('/card/listCard');
+            return;
+        }
         let result = await Card.destroy({ where: { id: card.id } });
         console.log(result + ' Payment Card deleted');
         res.redirect('/card/listCard');
